Move more dependency list computations over to the buildplan package.
